+++ demo/source/demos/codeSplitting/Root.js
@@ -6,18 +6,6 @@ import routes from "./routes";
 import Pages from "./Pages";
 
 class Root extends Component {
-    /** This will hold the dynamically loaded component */
-    state = { Page: null };
-
-    /**
-     * We handle the onChange callback from JARL because objects
-     * resolved during routing are available on the `resolved` key
-     */
-    onChange = ({ resolved }) => {
-        // Store it in state
-        this.setState({ Page: resolved.Page });
-    };
-
     render() {
         const { history, basePath } = this.props;
         return (
@@ -25,10 +13,8 @@ class Root extends Component {
                 history={history}
                 routes={routes}
                 basePath={basePath}
-                onChange={this.handleChange}
             >
-                {/* Pass it into our app to be rendered */}
-                <Pages Page={this.state.Page} />
+                <Pages />
             </StateProvider>
         );
     }