+++ demo/source/demos/codeSplitting/code.js
@@ -1,15 +1,10 @@
 import routes from "!!raw-loader!./routes";
-import Root from "!!raw-loader!./Root";
 import Pages from "!!raw-loader!./Pages";
 
 export default [
     {
         name: "routes.js",
         code: routes
-    },
-    {
-        name: "Root.js",
-        code: Root
     },
     {
         name: "Pages.js",